Modifying the script to use common scripts
