+++ .github/scripts/release.main.kts
@@ -1,0 +1,6 @@
+#!/usr/bin/env kotlin
+
+@file:Import("../../../scripts/npm-release.main.kts")
+@file:Import("../scripts/utils.main.kts")
+
+releasePlugins(getAllPluginsPath(), false)
