+++ .github/workflows/pull_request_verification.yml
@@ -10,30 +10,29 @@ jobs:
     runs-on: macos-latest
     if: ${{ !github.event.pull_request.draft }}
     strategy:
-      fail-fast: false
-      matrix:
-        xcode: ["15.3"]
+      fail-fast: true
     steps:
+      - name: Checkout sdk automation scripts
+        uses: actions/checkout@v3
+        with:
+          repository: moengage/sdk-automation-scripts
+          path: sdk-automation-scripts
+          token: ${{ secrets.SDK_BOT_ACCESS_TOKEN }}
+      - name: Automation script setup
+        uses: ./sdk-automation-scripts/actions/npm-repository-setup    
       - name: Checkout code
         uses: actions/checkout@v3
-      - name: Setup node
-        uses: actions/setup-node@v3
-        with:
-          node-version: 20.5.1
-          registry-url: 'https://registry.npmjs.org'
-      - name: Set up JDK
-        uses: actions/setup-java@v3
-        with:
-          distribution: 'zulu'
-          java-version: 17
-      - name: Set up gradle cache
-        uses: gradle/gradle-build-action@v2
-        with:
-          cache-read-only: ${{ github.ref != 'refs/heads/master' && github.ref != 'refs/heads/development'}}
+        with: 
+          path: source
+          fetch-depth: 0
+          fetch-tags: true
+          token: ${{ secrets.SDK_BOT_ACCESS_TOKEN }}
       - name: Setup script
+        working-directory: source
         run: |
-          chmod +x .github/scripts/common-utils.main.kts
+          chmod +x .github/scripts/utils.main.kts
           chmod +x .github/scripts/verify-plugins.main.kts
       - name: Verify
+        working-directory: source
         run: |
           kotlinc -script .github/scripts/verify-plugins.main.kts
